requirements.txt
os
shutil
sqlite3
base64
json
win32crypt
pycryptodomex
